Moving test to the same package as the tested class.
